rename transform -> native, don't export generated submodules
